test: update atlas-service tests
