+++ packages/atlas-service/src/main.spec.ts
@@ -1,11 +1,11 @@
 import Sinon from 'sinon';
 import { expect } from 'chai';
-import { AtlasService, getTrackingUserInfo, throwIfNotOk } from './main';
+import { AtlasService, throwIfNotOk } from './main';
+import * as util from './util';
 import { EventEmitter } from 'events';
 import { createSandboxFromDefaultPreferences } from 'compass-preferences-model';
 import type { PreferencesAccess } from 'compass-preferences-model';
 import type { AtlasUserConfigStore } from './user-config-store';
-import type { AtlasUserInfo } from './util';
 
 function getListenerCount(emitter: EventEmitter) {
   return emitter.eventNames().reduce((acc, name) => {
@@ -74,6 +74,12 @@ describe('AtlasServiceMain', function () {
 
   let preferences: PreferencesAccess;
 
+  let getTrackingUserInfoStub;
+
+  before(function () {
+    getTrackingUserInfoStub = sandbox.stub(util, 'getTrackingUserInfo');
+  });
+
   beforeEach(async function () {
     AtlasService['ipcMain'] = {
       handle: sandbox.stub(),
@@ -114,8 +120,15 @@ describe('AtlasServiceMain', function () {
     sandbox.resetHistory();
   });
 
+  after(function () {
+    sandbox.restore();
+  });
+
   describe('signIn', function () {
     it('should sign in using oidc plugin', async function () {
+      const atlasUid = 'abcdefgh';
+      getTrackingUserInfoStub.returns({ auid: atlasUid });
+
       const userInfo = await AtlasService.signIn();
       expect(
         mockOidcPlugin.mongoClientOptions.authMechanismProperties
@@ -124,6 +137,9 @@ describe('AtlasServiceMain', function () {
         // proper error message from oidc plugin in case of failed sign in
       ).to.have.been.calledTwice;
       expect(userInfo).to.have.property('sub', '1234');
+      expect(preferences.getPreferences().telemetryAtlasUserId).to.equal(
+        atlasUid
+      );
     });
 
     it('should debounce inflight sign in requests', async function () {
@@ -519,19 +535,6 @@ describe('AtlasServiceMain', function () {
           "Can't sign out if not signed in yet"
         );
       }
-    });
-  });
-
-  describe('getTrackingUserInfo', function () {
-    it('should return required tracking info from user info', function () {
-      expect(
-        getTrackingUserInfo({
-          sub: '1234',
-          primaryEmail: 'test@example.com',
-        } as AtlasUserInfo)
-      ).to.deep.eq({
-        auid: '03ac674216f3e15c761ee1a5e255f067953623c8b388b4459e13f978d7c846f4',
-      });
     });
   });
